+++ tests/CollectorStaticTest.php
@@ -5,17 +5,21 @@ namespace tests\olvlvl\ComposerAttributeCollector;
 use olvlvl\ComposerAttributeCollector\Collector;
 use olvlvl\ComposerAttributeCollector\Config;
 
-final class ReferenceTest extends TestAbstract
+/**
+ * Test {@link Collector} with {@link Config::STRATEGY_STATIC}.
+ */
+final class CollectorStaticTest extends TestAbstract
 {
     #[\Override]
     protected static function getStrategy(): string
     {
-        return Config::STRATEGY_REFERENCE;
+        return Config::STRATEGY_STATIC;
     }
 
+    #[\Override]
     protected static function dump(Config $config): void
     {
         $collector = new Collector($config, new FakeLogger());
-        $collector->dump();
+        $collector->run();
     }
 }