+++ tests/Factory/CollectionRendererFactoryTest.php
@@ -1,0 +1,45 @@
+<?php
+
+namespace tests\olvlvl\ComposerAttributeCollector\Factory;
+
+use olvlvl\ComposerAttributeCollector\CollectionRenderer\ReferenceCollectionRenderer;
+use olvlvl\ComposerAttributeCollector\CollectionRenderer\StaticCollectionRenderer;
+use olvlvl\ComposerAttributeCollector\Config;
+use olvlvl\ComposerAttributeCollector\Factory\CollectionRendererFactory;
+use PHPUnit\Framework\Attributes\DataProvider;
+use PHPUnit\Framework\TestCase;
+
+final class CollectionRendererFactoryTest extends TestCase
+{
+    #[DataProvider('provideRenderer')]
+    public function testFrom(string $strategy, string $expected): void
+    {
+        $config = self::makeConfig($strategy);
+
+        $this->assertSame($expected, CollectionRendererFactory::from($config));
+    }
+
+    /**
+     * @return array<string, array{ string, class-string }>
+     */
+    public static function provideRenderer(): array
+    {
+        return [
+            'static' => [ Config::STRATEGY_STATIC, StaticCollectionRenderer::class ],
+            'reference' => [ Config::STRATEGY_REFERENCE, ReferenceCollectionRenderer::class ],
+        ];
+    }
+
+    private static function makeConfig(string $strategy): Config
+    {
+        return new Config(
+            vendorDir: __DIR__,
+            attributesFile: __DIR__ . '/attributes.php',
+            include: [ __DIR__ ],
+            exclude: [],
+            useCache: false,
+            isDebug: false,
+            strategy: $strategy,
+        );
+    }
+}